+++ runbot.js
@@ -1,0 +1,1 @@
+require("./bp.js").runBot("tokerino here");
